Reworked the data access and removed three classes!!! Woohoo!

+++ Source/Momntz.UI/Core/TagHtmlHelper.cs
@@ -11,7 +11,7 @@ namespace Momntz.UI.Core
         {
             if(tags != null)
             {
-                List<string> tgs = tags.Where(t => t.Kind == KindOfTag.Tag)
+                List<string> tgs = tags.Where(t => t.Kind == (int)KindOfTag.Tag)
                     .Select(tag => string.Format("<a href=\"/{0}/Tag/{1}\" title=\"View all momentos in {1}\" rel=\"category tag\">{1}</a>", username, tag.Name))
                     .ToList();
